requirements.txt
bcrypt==3.2.0
mysql-connector-python==8.0.27
